backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
mysql-connector-python==8.3.0
python-dotenv==1.0.0
pydantic==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
